Update _loader.s
Changed # of page tables from 512 to 256 since kernel will only be mapped in the top 256 Mb (0xF0000000). Fixed page fault bug when jumping to kernel after enabling paging. This bug was caused due to applying the page table descriptor into the incorrect entry in the page table. The page table index (Which is bits 19:12 in the virtual address) must be multiplied by 4 before being concatenated to the page table base address!