Implemented _vmm_kernel_init, updated spinlocks, made PL1 stacks page aligned
The kernel's vmap structure is now fully initialized by the vmm system.
vmm_init can now be called to set up the vmm system.

Made the PL1 stacks page aligned in boot.s

Changed the spinlock structure to a typedef of an uint32_t since that is
essentially what a spin lock is.

Use spin_irqlock in kprintf