feat(kheap, iomem, kmain): Implemented kheap_stats
Implemented kheap_stats function which returns the bytes of memory
allocated and freed and the number of memory blocks allocated and freed.

Also added the offset for the IMSC (interrupt mask register) for UART0
in iomem.h for both realview and versatile platforms.

Added a rudimentary data abort exception handler