Fix(Makefile,_evt.s,pmap.c): Various fixes
Makefile:
objcopy now dumps the .bss section into the image as well. This makes
the image larger but it doesn't rely on the system to load the .bss
section for the kernel.

_evt.s:
We are overwriting the vector base address register with a new value,
why did I think it was a good idea to OR the existing value in the
register with a new value?

pmap.c:
Added the AP0 bit to page table entries. This prevents the MMU from
raising an Access Flag Fault when accessing those pages for the first
time.