Fix(putchar.c): Added carriage return after every newline
So putchar returns -1 if the UART TX FIFO buffer is full, but kputc wasn't
waiting until it wasn't full before sending the next character.
Essentially, if the buffer was full kputc would be skipping characters!
Not good. The easy fix? just have a while loop in kputc that waits until
the buffer is not full

Another problem: This may be just a thing with minicom but sending
newlines (\n) would just do that: a newline. There would be no implied
carriage return so newlines would start on the newline in the minicom
terminal but would not start at the beginning of said line. Modified
putchar (in the beaglebone platform code) to send a carriage return (\r)
after every newline