lock: Add exclusive/shared lock module
This lock allows multiple threads to have shared ownership of the lock and only one thread to have exclusive ownership
of the lock. It will sleep threads that can't gain ownership and will wake threads once the lock is released.
Currently, priority inversion is only handled if the lock is used with only exclusive ownership. There's no way to
handle shared owners and priority inversion.